전남대 BE_유보민 3주차 과제(1단계) (#261)
* Move from spring-wishlist

* fix : properties 추가

* fix : Model의 User에서 Entity의 Users를 사용하는 것으로 수정

* fix : Model의 User에서 Entity의 Users를 사용하는 것으로 수정

* fix : primitive 자료형으로 수정, Id 지정

* delete : JPA로 refactor하면서 사용하지 않게 된 클래스 삭제

* fix : Model의 Product에서 Entity의 Products를 사용하는 것으로 수정

* fix : 3주차 과제에 맞게 수정

* fix : ProductJpaRepository를 사용하는 것으로 수정

* fix : int -> long으로 수정

* fix : Model의 User에서 Entity의 Users를 사용하는 것으로 수정

* delete : 사용하지 않는 클래스 삭제

* fix : int -> long으로 수정

* fix : UserJpaRepository를 사용하는 것으로 수정

* fix : productId를 주소로 넘기지 않도록 수정

* fix : Pathvariable 삭제, Entity인 Users를 사용하는 것으로 수정

* fix : int -> long으로 수정, 생성자 추가

* fix : WishListJpaRepository를 사용하는 것으로 수정

* feat : ProductJpaRepository 작성

* feat : Products 엔티티 작성

* feat : UsersJpaRepository 인터페이스 작성

* feat : Users 엔티티 작성

* feat : WishListJpaRepository 작성

* feat : WishList 엔티티 작성

* feat : WishList 엔티티의 복합키 클래스 작성

* feat : JpaRepository에 대한 Test 클래스 작성

* delete : 사용하지 않는 클래스 삭제

* fix : 주석 삭제

+++ src/main/java/gift/Model/User.java
@@ -1,7 +1,7 @@
 package gift.Model;
 
 public class User {
-    private int id;
+    private long id;
     private String email;
     private String name;
     private String password;
@@ -11,7 +11,7 @@ public class User {
       
     }
 
-    public User(int id, String email, String password, String name, boolean isAdmin) {
+    public User(long id, String email, String password, String name, boolean isAdmin) {
         this.id = id;
         this.email = email;
         this.password = password;
@@ -42,12 +42,12 @@ public class User {
     public void setName(String name) {
         this.name = name;
     }
-
-    public int getId() {
+  
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
